# server-tv-pairing/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flint device pairing API",
    "version": "1.1.0",
    "description": "Encrypted VPN configuration handoff for Android, Android TV, Windows and iOS. No account session is transferred. Owner QR: HTTPS /connect/{id}#key=...; receiver QR: flint://pair?v=1&id=...#key=... . The production server has not been changed by this package."
  },
  "servers": [
    {
      "url": "https://flintmain.ru"
    }
  ],
  "paths": {
...
    "/api/v1/devices/pairing/download/android": {
      "get": {
        "operationId": "DownloadFlintAndroid",
        "security": [],
        "responses": {
          "200": {
            "description": "Public configured software download",
            "content": {
              "application/vnd.android.package-archive": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "HTTP byte range"
          },
          "404": {
            "description": "Distribution file is not configured"
          }
        }
      }
    },
    "/api/v1/devices/pairing/download/windows": {
      "get": {
        "operationId": "DownloadFlintWindows",
        "security": [],
        "responses": {
          "200": {
            "description": "Public configured software download",
            "content": {
              "application/vnd.microsoft.portable-executable": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "HTTP byte range"
          },
          "404": {
            "description": "Distribution file is not configured"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}